Require manual tests
